Adjust baseurl for pgdgNN-updates-debuginfo repository
The baseurl is constructed to use rhel-$releasever-$arch (as usual), but
the repository doesn't have directories on the server for rhel-7Server-*
for RHEL, only rhel-7.x-* for CentOS.

This was reported on pgsql-bugs:

https://www.postgresql.org/message-id/16307-68805138317a0bf0@postgresql.org

Since there has been no response (also to a followup by email), we have
no option but to munge baseurl according to ansible_distribution_version
so that we can install debuginfo packages once again. This block can be
removed if the repository is ever fixed.

References: 13944