Always ensure barman_home has the right ownership
Installing the barman package will normally do this, but if you include
barman in ``packages`` in order to use a pre-initdb hook to bring up an
instance from a restored backup, the package will be installed before
sys/volumes. On instances with a separate ``volume_for: barman_data``,
mounting a volume over barman_home makes it end up owned by root:root
instead of barman:barman.

Although this is not a common use-case, it's easy to fix with a single
added task that will usually be a no-op on most deployments.

References: 15490