Require $playbook to be the next argument after ``tpaexec playbook $cluster``
We now accept only the following invocations of tpaexec playbook:

    tpaexec playbook $cluster/$playbook …

    tpaexec playbook $cluster $playbook …

The latter form is supported so that you can invoke a playbook that's
outside the cluster directory by specifying a full path.

In the past, you could run ``tpaexec playbook $cluster … $playbook …``,
but we now require the playbook to be specified immediately after the
cluster directory. This simplifies the logging of actions performed
against a cluster.