Add ``platform: shared``
An instance can now specify ``platform: shared`` and ``hosted_by: xxx``,
where xxx is the name of an existing, non-shared (bare, aws, whatever)
instance. The shared instance is set up with the same IP addresses as
the instance it is hosted_by.

This is the initial provisioning support—deployment needs many changes
(apart from the obvious one of not applying the sys role to a shared
instance).

References: RM13961