Don't apply */pkg roles on shared instances
Instances decide what packages to install based on their assigned role.
We would prefer to let shared instances do the same, but we cannot rely
on the good nature of package managers to tolerate multiple concurrent
invocations on the same host.

The logic to determine if a shared instance x is hosted_by an instance y
that does not have the role(s) required to install all of the packages
that x needs is more complicated than justified by this odd use case.
(The main reason we want to support shared instances is to be able to
have more than one Postgres cluster on the same server.)

So we don't apply the pkg roles on shared instances, at the expense of
not being able to have an instance with role 'postgres' be able to host
a shared instance with role 'barman', say; the postgres instance would
not know it needed to install the barman package, and the shared one
wouldn't install anything.

A more palatable solution would be to introduce some sort of lock wait
on the package manager task. Yum already supports this natively—if you
run ``yum install -y x`` twice on the same server, one process acquires
the lock and installs the package, and in theory the others wake up and
realise that they don't need to do anything. In theory, this is exactly
what we need. In practice, (a) parallel yum invocations can break when
they need to download multiple dependency packages (e.g., yum install
openvpn), and (b) apt has no equivalent functionality.

It seems possible to implement this functionality robustly in tpaexec
(with flock(1), which we already use in, e.g., roles/secret) but given
that the problem is somewhat academic (given that shared instances are
mostly going to be n×postgres) and that there are potential portability
problems, it doesn't seem worth the effort.