Overhaul package installation in EC2 user-data
We were a bit lazy in terms of using image['name'] at template expansion
time to handle minor platform-level differences in setting up a python2
virtualenv to install awscli. Unfortunately, extending that to support
Ubuntu 20.04/Focal (where Python2 doesn't work) was a bit too clumsy.

Now we use /etc/os-release (as library/minimal_setup does) to detect the
target OS and decide what packages to install, and how. This means we no
longer need the "identify_os" filter hack, but we accept a few obscure
shortcuts in the interests of brevity.

We install awscli with Python3 where possible, and fall back to Python2
where necessary. In order to compensate for the extra code, we roll the
nvmecli and mdadm installation into the same process.