Move default failure timeout definitions
We want the default behaviour of postgres-down to not restart; but to be
able to restart by defining restart_after. At the same time, we want the
default partition failure to recover after some time. Different failures
can have different behaviour with regard to timeouts.

The CAMO crash tests all have similar timeout requirements, so we can
just set those in one place as a convenience.

References: 13859