Allow first_bdr_primary assignment to be overriden
Right now, we may assign first_bdr_primary to a newly-added instance,
which doesn't work with (i.e., completely ignores) an existing BDR node
group. This is a stopgap workaround to override this decision, e.g.,

    tpaexec deploy . -e first_bdr_primary=existing_node

This is just a hack, and I strongly recommend against setting
first_bdr_primary in config.yml.