Ensure first_bdr_primary is not a new instance
When re-deploying an existing BDR cluster with a new node added, we must
not designate the new node as first_bdr_primary, because it will happily
create a new BDR node group and ignore the existing one.

We still assign first_bdr_primary in roles/platform without worrying too
much about which node is selected, but now after running postgres/facts,
we check to see if any instances already have BDR active in the desired
bdr_database; and if first_bdr_primary is not one of these instances, we
override the earlier assignment.

It's no longer possible to support setting ``-e first_bdr_primary=…``,
because we may need to override that setting with set_fact; instead, we
now accept ``-e preferred_first_bdr_primary=…`` and use that to set the
initial value of first_bdr_primary, subject to later revision based on
the postgres/facts check results.

References: RM17254, RT67887