Avoid allocating a lot of memory to collect restorecon output
In RT68213, a customer reports that Ansible runs out of memory when
executing restorecon after they initialised PGDATA from a backup using a
pre-initdb hook; there are 2M+ files in PGDATA when restorecon runs.

Fortunately, we don't really need the output, we just need to know if
there was any output at all. So we can just pipe the output to wc -l and
see if the result was 0 or not.

At the same time, we can ignore ``semanage fcontext`` failures (which
are of the EEXISTS variety) and run restorecon unconditionally. This
makes the output handling more uniform (always 0 or not, never '').

References: 17741