Fix various inconsistencies with --overrides-from
The code to assign values in aws.process_arguments() (and other similar
functions) didn't take into account the possibility of values specified
through the --overrides-from mechanism.

In particular, it's now possible to specify ``locations`` in an override
file, which was not possible before. If you specify locations, you have
to provide a complete list of the required locations, and any missing
bits will be filled in (e.g., region, subnet, az).

References: 18013