Add multiple distribution support to platforms/docker
Until now, we used the centos/systemd image by default for Docker. You
could choose another image with ``--os-image``, but it was entirely up
to you to ensure that the image was systemd-enabled.

Since there are no other official systemd-enabled images, this meant
that it was extremely inconvenient to use anything but centos/systemd
(although it was certainly possible).

We now accept ``--os Debian/Ubuntu/RedHat`` (with the applicable choices
for ``--os-version``) and build whichever ``tpa/label:version`` image is
required during docker/provision. We include Dockerfiles to build images
for Debian 8/9/10, Ubuntu 16.04/18.04/20.04, and CentOS 7/8, along with
a distribution-independent overlay to configure systemd properly.

You can customise the image building process by specifying build
instructions via ``docker_images``:

    docker_images:
        "some/image:ver":
            dockerfile: Dockerfile
            context_path: /opt/docker-images/some-image-final

Use ``--overrides-from docker_images.yml`` or add this to the top level
of config.yml to specify a path to the custom Dockerfile to use to build
the given image name (which can be referenced via --os-image as usual).
See platforms/docker/docker_image.yml for more details on the two-stage
build process.

Idea and implementation by Craig, with some reorganisation by me.