Remove most ``tags: always`` from roles/sys
As a matter of caution, it's reasonable to tag all include tasks with
``always``, because we don't know whether the included tasks may have
some task that needs to always run (e.g., a set_fact).

In this case, however, only sys/volumes and sys/openvpn need to always
be included (and even that is something we can try to avoid, but not as
part of this particular easy change).