Add roles/harp to install and configure HARP
There's a new --enable-harp configure option that sets
cluster_vars.enable_harp. This causes the HARP package to be installed
on bdr+primary instances, and the harp_router/harp_manager services to
be started. Also, haproxy is configured to perform status checks to
harp_manager instead of the regular backend queue checks.