Accept user-defined entries in postgres_users
It is now convenient to specify Postgres users to create:

    cluster_vars:
      postgres_users:
      - username: example
      - username: xyzzy
        generate_password: false
        role_attrs:
          - superuser
          - replication

These users are created in postgres/config/final with tpaexec-generated
database users such as repmgr and barman.

It was always possible to specify entries in postgres_users, because
postgres/createuser would just append new users to the list, but the
password handling made it inconvenient to use. With this change, we
generate passwords for user-defined entries (unless suppressed by
setting ``generate_password: false``).