Revise tagging in roles/sys/tasks/main.yml
We now use the following pattern:

    - include…
        name: …
        apply:
          tags: x
      tags: x

Note that the invocation of roles/sys from deploy.yml follows a similar
pattern, so everything is tagged with "sys" already.

Now ``--tags sys`` includes everything, and ``--skip-tags sys`` excludes
as much as it can (with the exception of openvpn, which is tagged always
because of the openvpn/ip tasks; these will move to roles/platforms in
future). Also, ``--tags pkg`` (for example) will execute only the pkg
tagged tasks in sys/pkg.

We retain some tags within sys that we probably don't need and don't
particularly want to support (e.g., sysstat, rsyslog) just to avoid
breaking compatibility for now. This is subject to future revision.