Add support for user-defined databases in ``postgres_databases``
You can now specify databases to create during deployment:

    postgres_databases:
    - name: dbname
      owner: somerole
      encoding: xyzzy
      lc_collate: x
      lc_ctype: y
      template: template1
    - name: otherdb
      …

All of the parameters other than the database name are optional.

We also use this functionality to create the bdr_database, instead of
doing it separately in postgres/bdr. This allows the user to specify a
custom configuration for bdr_database in config.yml. Otherwise the code
uses the current defaults to create it.

Note that postgres_databases entries can specify ``template: template0``
explicitly in order to avoid inheriting the extensions created by
default.