Exclude 'readonly' instances from setting allow_lead=1
BDR instances may have role bdr+primary, or bdr+primary+readonly (for
logical standby instances), or bdr+replica (for physical replicas of a
BDR instance). Only the first of these should be allowed to be marked
as the leader of a cluster.

We can reuse first_bdr_primary_candidates to make this decision, rather
than ``role|contains('bdr','primary') and not 'readonly' in role`` or
similar. This gives us an added element of control (in terms of being
able to override the assignment of first_bdr_primary_candidates) over
the instances where we configure HARP.