Improve pglogical provider_dsn derivation
This commit makes a few significant changes:

1. Under ``publications``, every entry must now specify a ``name``;
   there is no default based on the hostname and database name. The
   given name will be used for the pglogical node name, and also to
   reference the publication from ``subscriptions``.

2. As before, every entry under ``subscriptions`` must specify either
   the .provider_dsn or .publication.name; but it is now optional to
   specify .publication.instance. If you use unique names across the
   cluster for publications, you can reference publications by name
   alone and leave out the instance name altogether.

It was difficult enough to implement the "explicit .provider_dsn or
publication reference" logic correctly in Jinja2 earlier, even before
making the instance name optional (the problem being to suppress the
evaluation of .publication.* when provider_dsn is defined). It's much
easier to do it all in a new |provider_dsn filter.

(We keep some redundant sanity-checking asserts in subscribe.yml because
they produce nicer error messages than the checks in the filter.)