Avoid rewriting shared_preload_libraries once per library
We now handle shared_preload_libraries as a list, add the extensions we
want based on postgres_extensions, add repmgr when required (a change to
the earlier logic), and merge the contents into shared_preload_libraries
on disk in one step.

As a side-effect of the above, the repmgr/{primary,witness} roles need
not include preload-library to add repmgr to shared_preload_libraries.

And as a side-effect of *that*, we can get rid of the
repmgr/{primary,replica,witness} roles altogether. Since these roles
have remained unchanged for a long time, we can conclude that there is
no other compelling need for them.

(Note: the repmgr/final/{primary,replica,witness} roles are still
needed, and remain unchanged here.)