Start repmgrd only after slot reconciliation
Because of the way we set up cascading replicas, Postgres may be started
on all instances before the necessary replication slots are configured.
If we start repmgrd at that point, it may refuse to start because the
repmgr.nodes table hasn't replicated yet:

    [NOTICE] using provided configuration file "/etc/repmgr/11/repmgr.conf"
    [NOTICE] repmgrd (repmgrd 5.1.0) starting up
    [INFO] connecting to database "host=…"
    [ERROR] no metadata record found for this node - terminating
    [HINT] check that 'repmgr (primary|standby) register' was executed for this node
    [INFO] repmgrd terminating...

Now we start repmgrd only after we fix up the slots. (That's not a
guarantee that replication will have caught up by the time repmgrd
starts, but it's better than a guarantee that it won't. We might
have to do some --wait-sync like checking… or perhaps repmgrd
should loop a little before giving up?)