Migrate from pgbouncer.get_auth to pg_catalog.pgbouncer_get_auth
Creating a separate pgbouncer schema was polite behaviour, but it caused
problems with BDR in various circumstances. We had an extra cleanup step
before the join that would drop the schema on all but one instance, and
then let it be synced up by BDR, but this did not solve all problems.

Impolite though creating a function in pg_catalog may seem, it seems to
be the least intrusive alternative. During a fresh redeploy, we rewrite
pgbouncer.ini to use the new auth_query in pgbouncer/config, create the
function in postgres/config/final, and let pgbouncer be restarted with
the changed configuration in pgbouncer/final.

Only then is it safe to drop the old function, which we do in a new
postgres/cleanup role. Here we are careful to drop the function first,
and then the schema without cascade, on the off-chance that someone was
inspired to create other objects in this schema.