Don't fetch pglogical.replication_set_table.set_row_filter
(a) pglogical2 doesn't have pglogical.get_row_filter_expr(), and there's
    absolutely nothing useful we can do with the raw set_row_filter
    value anyway, and

(b) we weren't doing anything useful with the row filter expression.

If we need it in future, we can add pglogical_version_num-specific
handling for the field (e.g., fetch a NULL on v2 and the expr on v3).