Collect per-database pg_extensions contents in cluster_discovery
Use this to skip invoking postgresql_ext when the extension exists. It's
easy to do this for extensions because they can only exist or not, they
aren't going to be altered by postgresql_ext once they are created.

To do something similar for users or databases, we would have to not
only test if they exist, but also compare their attributes to see if
the postgresql_{db,user} invocation might alter them.