Add a ``preload_extensions`` setting
If you are installing extensions that (a) require an entry in
shared_preload_libraries, and (b) are not already mentioned in
default_preload_extensions (roles/postgres/config/defaults), you
can add its name to the ``preload_extensions`` list.

This does nothing more than declare that an extension needs an entry in
shared_preload_libraries *if it is present*. So to actually install the
extension, you must mention it in extra_postgres_extensions as always.