Empty default_unwanted_packages list
We listed some packages in there including nano that were removed for
"historical reasons", and don't seem especially justified in light of
customer complaints about their presence.

Anything added back in here should be added because of a specific,
documented problem that it causes with the operation of a TPAexec
cluster.

References: RT68648