Make haproxy log to syslog
Change haproxy's configuration so that it logs to a socket not to
UDP; add a configuration file for rsyslog so that it reads from
this socket and writes the logs to their own file. On Debian, a
similar rsyslog configuration file may have been put in place by
the haproxy package, but we overwrite it with one that has more
cross-version-compatible syntax.

References: RM7032