Speed up sysctl settings
Instead of invoking the sysctl module in a loop to run sysctl(8) AND
edit /etc/sysctl.conf AND run sysctl -p, we can just use linesinfile to
add the desired key=value settings to /etc/sysctl.conf, and just call
sysctl -p once if the file was changed.

This depends on a new dict_format filter that returns a list consisting
of lines produced by applying a format_string (e.g., "{key}={value}") to
each key and value in the given dict.