Add support for setting password_encryption='scram-sha-256'
If you set password_encryption, the specified value is used:

1. It's set in postgresql.conf
2. It's set as the last field in pg_hba.conf
3. It's used to generate passwords for CREATE USER commands
4. It's used to configure pgbouncer's auth_method and userlist.txt

If you don't set password_encryption, the default for existing clusters
continues to be whatever value they are using, which is probably 'md5'.
For new clusters running Postgres 10+, the default is 'scram-sha-256'.

This adds 'passlib' to python-requirements.txt

References: RM17625