Strip leading ../* when building PDF links
We cat `configure/xxx.md` into tpaexec.md, but the links within that
file are to things like `../tpaexec-hooks.md`, which won't work. If
we just strip the leading ..s, it will be fine. Links to internal pages
from top-level pages of the form `configure/xxx.md` will continue to
work, because single-pdf.py will be able to open them.