pgbouncer: add SCRAM support, separate pgbouncer{,_auth_user}
We used a user named pgbouncer both as the user with which to connect to
the pgbouncer administration pseudo-database, and as the auth_user which
pgbouncer itself uses to connect to postgres and execute the auth_query.

Conflating these two users was unnecessary and, in the context of SCRAM,
rather unhelpful. We can't just hand pgbouncer a SCRAM-hashed password
(as we do with `CREATE USER`) and expect it to be able to authenticate
to Postgres—it must either have access to the plaintext password to use
SCRAM, or continue to use MD5. Obviously, the latter is preferable.

1. We now create the `pgbouncer_auth_user` user with an MD5 password in
   pgbouncer/config, add an entry for it in userlist.txt, use it as the
   auth_user with auth_type=md5, and allow pgbouncer to authenticate as
   that user in pg_hba.conf.

2. We no longer create the pgbouncer user in the database, unless it's
   explicitly included in postgres_users. We still generate a password
   for it, so that we can create an entry for it in userlist.txt. This
   user is now used only for pgbouncer administration.

3. If the pgbouncer user exists in the database (and isn't mentioned in
   postgres_users), we revoke its permissions and drop it.

4. If the pg_catalog.pgbouncer_get_auth() function doesn't exist, or
   pgbouncer_auth_user doesn't have execute permissions in it, we now
   run CREATE FUNCTION/GRANT in postgres/config/final.

You can still set `pgbouncer_auth_user: pgbouncer` to use the same user,
but there's really no advantage to doing so.