Revise known_hosts generation
We used to replace /etc/ssh/ssh_known_hosts on target instances with a
template that generated entries for every generated public host key and
every hostname/address of every instance in the cluster, regardless of
whether that instance had the generated host keys installed or not.

1. We now split sys/hostkeys away from sys/hosts, and give both unique
   tags (hosts and hostkeys, respectively).

2. We now install the generated hostkeys on (a) Docker containers (as
   before), and (b) bare instances with `manage_ssh_hostkeys` enabled.
   We assume that other kinds of instances will install the host keys
   themselves (as EC2 instances do in user-data, and Vagrant instances
   do in the Vagrantfile).

3. We use a cut-down known_hosts.j2 in sys/hostkeys to do an in-memory
   cross-join between the list of all hostnames, aliases, and addresses
   and the corresponding public keys (either the ones we generated, or
   the ones we fetched from instances with `manage_ssh_hostkeys` off),
   and hand the resulting entries to linesinfile to update known_hosts.

4. We introduce two new filters to aid in the above: `extract_keys` and
   `ternary_format`.

5. By default, we set `manage_ssh_hostkeys: no` for bare instances.

The above is all during deployment. There's one change to the generation
of tpa_known_hosts at provision time, which is to include entries (which
are incorrect at the time of generation) for bare instances that do have
manage_ssh_hostkeys enabled, because they will be needed once we replace
the hostkeys during deployment.

The upshot is that we no longer overwrite ssh_known_hosts, and we now
install correct entries for instances with `manage_ssh_hostkeys` off.

References: RM6060