Look for `python2.7` on Ubuntu
As in commit #f4f7196 for Debian 8, we extend minimal_setup to return
`python=python2.7` for Ubuntu, because there's only /usr/bin/python2.7,
no symlink to python2. At the same time, we also fix the condition to
check for `.startswith('python2')`, so that it would not try to install
python3 elsewhere if the python2 link *did* exist.

References: RM19184