Allow hba for pgbouncer_auth_user on localhost with md5
If an instance had roles postgres+pgbouncer, pgbouncer was not able to
connect to Postgres because (a) the 127.0.0.1/32 entry matches first,
but requires scram-sha-256 authentication, and (b) the expected entry
uses a different IP address anyway.

We add an entry for pgbouncer_auth_user earlier in the file to avoid
this problem.

References: RM19379