Exclude : from generated passwords
Having : in generated passwords doesn't cause any problem in tpaexec
itself, because postgres/pgpass knows to escape \ and : in passwords
before writing .pgpass files.

However, excluding it to avoid confusing matters in other contexts, as
we do for \ and ", seems wise because writing unescaped : to .pgpass
could conceivably cause a security problem.