Increase wait for running pg_receivewal
Marc reports that it "sometimes" takes longer than the expected 15s.
This surprises me, but in the absence of concrete analysis, it doesn't
hurt to extend the timeout a little. If this ever happens again, though,
the answer is probably not to extend the timeout further, but to figure
out why the pg_receivewal startup is delayed and make it happen faster.