Specify labels to match when deprovisioning docker containers
Since the docker inventory plugin matches only the container name (and
not the Cluster=xxx tag, as we do for ec2.ini), we now specify labels
that the container must have before removing it.

This depends on feature-2.8/docker-absent-labels in 2ndQuadrant/ansible,
which makes docker_container match any specified labels before removing
a container.

References: RM19950

Signed-off-by: Haroon <muhammad.haroon@2ndquadrant.com>