Allow setting postgresql_flavour=epas via --epas configure option
This comprises the following pieces:

1. Add --epas configure option to set postgresql_flavour
2. Add EPAS entry for --install-from-source (untested)
3. Configure EDB apt/yum repository and authentication via
   EDB_REPO_CREDENTIALS_FILE containing "username:password" obtained
   from https://repo.enterprisedb.com
4. Configure separate epas_xxx_packages lists with the correct package
   names (there are no automatically-installable debuginfo packages for
   EPAS, they must be downloaded from https://builds.enterprisedb.com)
5. Fix default_postgres_service_names, default_postgres_bin_dirs, the
   path to the postmaster binary, etc.

Right now we don't adopt any of the EPAS conventions (port 5444 etc.).
We can sort all of that out later.

This branch is now able to set up a BDR-Simple cluster on RH8 with
EPASv12 and pglogical/BDR from ci-spool/*/snapshot, and I'm merging it
so that people can start testing it. Debian doesn't work at the moment,
but I'll investigate and fix that separately.

Note that not all combinations of versions of each package are available
on all distributions (e.g., there's no EPASv11 for Ubuntu 20.04/focal).