Use a separate task to ensure that postgresql_flavour is defined
For some reason related to the semantics of Ansible, setting a fact in
a task that has a "vars" clause does not allow an undefined variable
to occur in any of the expressions setting the fact, even if that
variable has the |default() operator applied.

The fix is more a workaround, in the sense that we just set that fact
in a separate task that doesn't need the "vars" clause.

Fixes RM20174.

Signed-off-by: Gianni Ciolli <gianni.ciolli@enterprisedb.com>