Set unix_socket_directories according to postgres_family
If we're installing EPAS, the client libraries default to trying to
connect to a socket in /var/run/edb-as. We can either set PGHOST in
target_environment to force the use of a specific directory, or adapt
the defaults gracefully as in this commit.

Also overhaul the handling in postgres/user of the run directories and
their tmpfiles.d entries (though we expect only one entry in practice).

Note: we used to set unix_socket_directories to always include both
/var/run/postgresql and /tmp regardless of distribution. We retain this
behaviour for compatibility with existing clusters, but don't continue
it with EPAS (there doesn't seem to be a compelling reason to do so).