Align postgres/enterprisedb local lines properly in pg_hba.conf
This reverts commit b9cb731 ("Use postgres_user in pg_hba.conf instead
of hardcoding postgres") and uses postgres/enterprisedb/other blocks
that preserve the correct alignment of "peer" at the end of the line
for the two most common values of postgres_user.

This is mostly cosmetic, but it avoids changing pg_hba.conf on existing
systems. If you use a non-standard postgres_user, it will no longer try
to align the "peer" at all.

(Fixed-width columns in configuration files, such a charming idea.)