Add Dockerfile and instructions
We support running tpaexec directly on OSX, but there are no packages,
and setting up the required dependencies has become more difficult. We
now suggest that MacOS X users use this Dockerfile to get started, and
run tpaexec in a container with a known-good environment, while still
being able to use it to create Docker containers.

This is a squashed commit of the dev/rm20037 branch by Haroon, with a
review-by-following-instructions from Simon M.

References: RM20037

Signed-off-by: Haroon <muhammad.haroon@2ndquadrant.com>
Signed-off-by: Abhijit Menon-Sen <ams@2ndQuadrant.com>