Don't say "unknown platform" if platform.xxx exists but can't be imported
To generate a cluster configuration with `--platform xxx`,
Platform.load() would try to import tpaexec.platform.xxx, and fail with
"unknown platform" if it caught an ImportError.

If platform/xxx.py exists but the import fails for some reason, say a
syntax error, Platform.load would fail with a backtrace, which is what
we want. However, if the import fails because xxx.py tries to import a
module that does not exist, it raises ModuleNotFoundError, which is a
subclass of ImportError, and therefore matches the `except ImportError`
in the caller, and misleadingly reports an unknown platform.

This was originally reported by Craig while he was working on docker.py,
so it's a problem with a very narrow scope. Still, we do want to be able
to distinguish between "unknown platform" (most likely a typo in the
`tpaexec configure --platform …` argument) and "we could not import
docker.py" (say), so we use importlib's util.find_spec() to test if
the module exists before actually importing it.

https://airbrake.io/blog/python/importerror-and-modulenotfounderror has
more details about these exceptions.

https://stackoverflow.com/questions/14050281/how-to-check-if-a-python-module-exists-without-importing-it
has more details about importlib.

Closes #20