Set `my_hosts_lines: []` when calling terminate-for-rehydration
We depend on roles/facts being executed for every instance, but we do
invoke terminate-for-rehydration.yml with --limit set to the names of
the instances to rehydrate. Rather than trying to change that logic to
include more instances (risky), we just ensure that my_hosts_lines is
set to [] so that the task in roles/facts/tasks/sys.yml to aggregate
them does not break with an error like this:

2021-02-18 18:22:51,731 p=7932 u=xqs678 |  fatal: [tpa-dba-test-de-barman]: FAILED! => {
    "msg": "The task includes an option with an undefined variable. The error was: 'ansible.vars.hostvars.HostVarsVars object' has no attribute 'my_hosts_lines'\n\nThe error appears to be in '/opt/2ndQuadrant/TPA/roles/facts/tasks/sys.yml': line 28, column 3, but may\nbe elsewhere in the file depending on the exact syntax problem.\n\nThe offending line appears to be:\n\n\n- name: Aggregate /etc/hosts lines across hosts\n  ^ here\n"
}

It's appropriate to do this here to emphasise that it's a special case
in invoking terminate-for-rehydration (which will never use the value
computed), rather than change the task to assume an empty default or
something like that.