Embed EDB credentials in sources.list.d entry on Debian 9/stretch
The /etc/apt/auth.conf.d mechanism for authentication was introduced in
Debian 10/buster, so we have to include the username and password in the
repository URL for earlier versions. (However, even Ubuntu 16.04 has an
auth.conf.d, so this is a Debian-only consideration.)

If we're installing the EDB repository entry, we make sure to do so in a
file named edb-$release.list, rather than the apt_repository module's
default name (which we remove, for compatibility with old clusters,
although EPAS support is not yet released).

Also ensure we don't log any credentials in such cases.

Closes #128