Allow BDR instances to be rehydrated
We formerly refused to terminate a primary instance for rehydration, but
this check is no longer applicable to BDR instances (now that we can use
pool-disable-server and pool-enable-server around a rehydration).

We do not check that the instance has been removed from the haproxy pool
before termination, however. Caveat emptor. (Of course, haproxy's health
checks would see the terminated server and behave sensibly afterwards.)