Avoid undefined errors during terminate-for-rehydration
The expression we used to calculate pgbouncer_max_client_conn tried to
take into account various possibilities, but it did not account for the
possibility that `|map('extract',hostvars,'haproxy_maxconn')` would give
us a list of nulls.

So if you tried to rehydrate a single pgbouncer instance, this task
would blow up with an error like this:

    «The task include an option with an undefined variable. The error
    was: None is undefined»

We rearrange the expression slightly to avoid this. (Note: we won't
calculate the same value as we would normally, but the actual value
doesn't matter, since we're going to terminate the instance anyway.)