Operate on haproxy+pgbouncer instances in parallel in disable-backend
If a BDR instance is a backend for multiple haproxy instances, we now
mark it as disabled on all applicable haproxies in parallel, and then
issue RECONNECT+WAIT_CLOSE in parallel to all applicable pgbouncers,
then call bdr.wait_slot_confirm_lsn(NULL, NULL) once at the end.

Although the earlier commit considered this a "future optimisation",
Vibhor convinced me not to overlook the operational impact of having
multiple delays piled one atop the other, and waiting for replication
to "complete" while there were potentially still active transactions.

After this change, we might still potentially wait a long time if any
active transactions refuse to exit… but at least we'll do this only in
parallel. If you don't want to wait at all, run

    tpaexec pool-disable-server . --nowait

This will issue RECONNECT, but not WAIT_CLOSE, to pgbouncer, with the
result that the command will not wait for active sessions to be closed.
Depending on what you are doing, you can terminate the sessions by hand
afterwards if you want, or leave them alone.