Make `tpaexec relink` add links to new commands and tests
If $TPA_DIR/architectures/{commands,tests} contain any files for which
an existing cluster directory does not have links (either because they
were removed, or because the cluster configuration predates the files),
relink will now create links to them.

If it finds broken symlinks that it can't repair, it removes them
instead of printing an error.

If deploy.yml is missing, relink will fix that too.

(Committed with minor editorialisation by AMS.)