Run cluster_discovery after starting Postgres
Suppose you run switchover in an M1 cluster and rehydrate the former
primary (now replica). When it starts up again, Postgres isn't running,
so we don't run cluster_discovery after minimal_postgres_setup, and we
don't update the role to 'replica'. Then we start Postgres, which sees
recovery.conf and starts streaming, but we still think it's a primary,
and postgres/config/final tries to create users, which fails.

The fix is to run cluster_discovery if we just started Postgres, and
therefore weren't able to collect facts during init.