Introduce a `failover_manager: [repmgr|harp]` setting
Only if you set `failover_manager: repmgr` (still the default) do we add
repmgr.so to shared_preload_libraries and enabled repmgr.service. We no
longer need to sprinkle "not 'bdr' in role" conditions here and there,
we can disable repmgr for BDR in one obvious place (roles/facts).

This supersedes the `enable_harp: yes` setting, which is still accepted
for backwards compatibility.