Install edb-asNN-server to get pg_receivewal in barman/pkg
As reported in https://enterprisedb.atlassian.net/browse/DB-1010, the
EPAS rpm client package does not contain pg_receivewal (although the deb
package does), and pg_receivewal is actually in the server package.

Until this is fixed, we need a workaround—after installing the client
package, we check if pg_receivewal is present, and if not install the
EPAS server package (only for EPAS, only on RedHat). If the problem is
fixed upstream, then pg_receivewal will already exist, and the server
package will not need to be installed.

Depending on the fix that is eventually committed upstream, we can
either retain this workaround with a version check, or perhaps even
remove it altogether.