Delay synchronous_standby_names setting on replicas too
I forgot that postgres/cleanup was not applied on replicas, so the
earlier change to delay activating synchronous_standby_names until
cleanup worked only for the primary.

Here we push the "'replica' not in role" condition down into the
pgbouncer change, which moves into a separate file; whereas the
sync-replicas.yml is applied to all (non-BDR) postgres instances.