Create efm user/group in postgres/user if needed
If we need to add postgres_user to the efm group (as we do in 586c54a),
we can't wait for efm/pkg to create the group much later. Oops. We must
create both group and user, because the efm post-install scripts don't
create the user if the group already exists.

(While we're at it, we also make sure that postgres_user's primary group
remains postgres_group, and any extra groups are declared separately.)