Fix systemctl disable/mask logic in common-systemd
The intention of the earlier code was to fetch a list of all units that
exist, and then disable the ones that we don't want one by one, because
systemctl complains if you try to disable a unit that doesn't exist.

However, the filtering of list-unit-files output was broken, and
installedunits always seemed to be set to a single empty line. Also, the
newline handling in UNITS_TO_MASK was broken, and we'd end up trying to
mask units with names like like foo.bar\x0a.

We can avoid the contortions: list-unit-files takes arguments, so we can
pass in the units we don't want, and disable the ones that are listed in
the output in a single command (i.e., no need to loop over the results).