Read 'installedunits' as an array to make shellcheck happy
Instead of treating the output from `systemctl list-unit-files` as a
string and relying on argument-splitting, we can read the lines into an
array and test/expand the arguments directly.

Since this command will be run only inside a Linux container, we don't
need to worry about the weird old bash version shipped with MacOS X—in
fact, we could even use mapfile -t here. Such freedom! So overwhelming!