Display help text from tpaexec custom commands
If there's an executable, we can invoke it with --help. If there's a
shell script and it defines a _tpaexec_help function, we can call it.
(If it's a playbook, though, there's nothing we can do to show help.)

We duplicate a part of try_as_playbook_or_script inside help() to do
this, but it's small enough that it's not worth trying to avoid the
duplication.

This commit also moves the help for relink into commands/relink, and
adds a _tpaexec_help function for M1/commands/switchover.sh as a
demonstration.

This makes it possible to have "first-class" custom commands with their
own help text without any changes to bin/tpaexec. We could move things
like store-password/show-password and rehydrate into custom commands.